Added all existing versions of the MiBand
This commit adds all hw and fw revisions to the firmware editor. Though not every hw revision seems to be getting the same icons, there will be some deeper digging in the future. It also adds a small tool i created, the "patternsearcher", which made it much faster to find the ~950 different patterns.